+++ L2J_Mobius_C4_ScionsOfDestiny/java/org/l2jmobius/gameserver/model/holders/ItemHolder.java
@@ -23,9 +23,9 @@ package org.l2jmobius.gameserver.model.holders;
 public class ItemHolder
 {
 	private final int _id;
-	private final long _count;
+	private final int _count;
 	
-	public ItemHolder(int id, long count)
+	public ItemHolder(int id, int count)
 	{
 		_id = id;
 		_count = count;
@@ -42,7 +42,7 @@ public class ItemHolder
 	/**
 	 * @return the count of items contained in this object
 	 */
-	public long getCount()
+	public int getCount()
 	{
 		return _count;
 	}
